+++ crates/chart/src/drawing/primitives_v2/brushes/highlighter.rs
@@ -120,7 +120,7 @@ impl Primitive for Highlighter {
                 let x = ctx.bar_to_x(bar);
                 let y = ctx.price_to_y(price);
                 ctx.begin_path();
-                ctx.arc(x, y, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+                ctx.arc(x, y, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
                 ctx.fill();
                 ctx.stroke();
             }
