+++ crates/chart/src/drawing/primitives_v2/projection/projection.rs
@@ -125,13 +125,13 @@ impl Primitive for Projection {
 
             // Point 1
             ctx.begin_path();
-            ctx.arc(x1, y1, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+            ctx.arc(x1, y1, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
             ctx.fill();
             ctx.stroke();
 
             // Point 2
             ctx.begin_path();
-            ctx.arc(x2, y2, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+            ctx.arc(x2, y2, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
             ctx.fill();
             ctx.stroke();
         }
