+++ crates/chart/src/drawing/primitives_v2/traits.rs
@@ -731,7 +731,7 @@ pub fn draw_control_points(ctx: &mut dyn RenderContext, screen_points: &[(f64, f
 
     for (x, y) in screen_points {
         ctx.begin_path();
-        ctx.arc(*x, *y, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+        ctx.arc(*x, *y, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
         ctx.fill();
         ctx.stroke();
     }
